requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
numpy==2.1.2
openai==1.51.2
pgvector==0.3.5
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tqdm==4.66.5
typing_extensions==4.12.2
uvicorn==0.31.1
aiohttp==3.9.1
